requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
django-cors-headers==4.3.1
django-filter==25.1
faker==37.3.0
